Add Rerun Action to PipelineRunProvider
