+++ src/components/PipelineRun/components/RerunPipelineButton.tsx
@@ -1,17 +1,13 @@
 import { useNavigate } from "@tanstack/react-router";
 import { RefreshCcw } from "lucide-react";
-import { useCallback, useRef, useState } from "react";
+import { useCallback } from "react";
 
 import TooltipButton from "@/components/shared/Buttons/TooltipButton";
-import { isAuthorizationRequired } from "@/components/shared/GitHubAuth/helpers";
-import { useAuthLocalStorage } from "@/components/shared/GitHubAuth/useAuthLocalStorage";
-import { useAwaitAuthorization } from "@/components/shared/GitHubAuth/useAwaitAuthorization";
 import useToastNotification from "@/hooks/useToastNotification";
-import { useBackend } from "@/providers/BackendProvider";
+import { usePipelineRun } from "@/providers/PipelineRunProvider";
 import { APP_ROUTES } from "@/routes/router";
 import type { PipelineRun } from "@/types/pipelineRun";
 import type { ComponentSpec } from "@/utils/componentSpec";
-import { submitPipelineRun } from "@/utils/submitPipeline";
 
 type RerunPipelineButtonProps = {
   componentSpec: ComponentSpec;
@@ -20,16 +16,9 @@ type RerunPipelineButtonProps = {
 export const RerunPipelineButton = ({
   componentSpec,
 }: RerunPipelineButtonProps) => {
-  const { backendUrl } = useBackend();
+  const { rerun, isSubmitting } = usePipelineRun();
   const navigate = useNavigate();
   const notify = useToastNotification();
-
-  const { awaitAuthorization, isAuthorized } = useAwaitAuthorization();
-  const { getToken } = useAuthLocalStorage();
-
-  const [isSubmitting, setIsSubmitting] = useState(false);
-
-  const authorizationToken = useRef<string | undefined>(getToken());
 
   const onSuccess = useCallback((response: PipelineRun) => {
     navigate({ to: `${APP_ROUTES.RUNS}/${response.root_execution_id}` });
@@ -43,34 +32,9 @@ export const RerunPipelineButton = ({
     [notify],
   );
 
-  const handleRerun = useCallback(async () => {
-    setIsSubmitting(true);
-    const authorizationRequired = isAuthorizationRequired();
-    if (authorizationRequired && !isAuthorized) {
-      const token = await awaitAuthorization();
-      if (token) {
-        authorizationToken.current = token;
-      }
-    }
-    await submitPipelineRun(componentSpec, backendUrl, {
-      authorizationToken: authorizationToken.current,
-      onSuccess: (data) => {
-        setIsSubmitting(false);
-        onSuccess(data);
-      },
-      onError: (error) => {
-        setIsSubmitting(false);
-        onError(error);
-      },
-    });
-  }, [
-    isAuthorized,
-    awaitAuthorization,
-    backendUrl,
-    componentSpec,
-    onSuccess,
-    onError,
-  ]);
+  const handleRerun = useCallback(() => {
+    rerun(componentSpec, { onSuccess, onError });
+  }, [componentSpec, rerun, onSuccess, onError]);
 
   return (
     <TooltipButton
