+++ src/providers/PipelineRunProvider.tsx
@@ -3,6 +3,7 @@ import {
   useCallback,
   useEffect,
   useMemo,
+  useRef,
   useState,
 } from "react";
 
@@ -10,6 +11,9 @@ import type {
   GetExecutionInfoResponse,
   GetGraphExecutionStateResponse,
 } from "@/api/types.gen";
+import { isAuthorizationRequired } from "@/components/shared/GitHubAuth/helpers";
+import { useAuthLocalStorage } from "@/components/shared/GitHubAuth/useAuthLocalStorage";
+import { useAwaitAuthorization } from "@/components/shared/GitHubAuth/useAwaitAuthorization";
 import {
   createRequiredContext,
   useRequiredContext,
@@ -23,6 +27,8 @@ import {
 } from "@/services/executionService";
 import { fetchPipelineRunById } from "@/services/pipelineRunService";
 import type { PipelineRun } from "@/types/pipelineRun";
+import type { ComponentSpec } from "@/utils/componentSpec";
+import { submitPipelineRun } from "@/utils/submitPipeline";
 
 type PipelineRunContextType = {
   details: GetExecutionInfoResponse | undefined;
@@ -31,7 +37,16 @@ type PipelineRunContextType = {
   status: string;
 
   isLoading: boolean;
+  isSubmitting: boolean;
   error: Error | null;
+
+  rerun: (
+    componentSpec: ComponentSpec,
+    options?: {
+      onSuccess?: (data: PipelineRun) => void;
+      onError?: (error: Error | string) => void;
+    },
+  ) => Promise<void>;
 };
 
 const PipelineRunContext = createRequiredContext<PipelineRunContextType>(
@@ -47,11 +62,17 @@ export const PipelineRunProvider = ({
 }) => {
   const { backendUrl } = useBackend();
 
+  const { awaitAuthorization, isAuthorized } = useAwaitAuthorization();
+  const { getToken } = useAuthLocalStorage();
+
   const [metadata, setMetadata] = useState<PipelineRun | null>(null);
   const [status, setStatus] = useState<string>("UNKNOWN");
 
+  const [isSubmitting, setIsSubmitting] = useState(false);
   const [isPolling, setIsPolling] = useState(true);
   const [error, setError] = useState<Error | null>(null);
+
+  const authorizationToken = useRef<string | undefined>(getToken());
 
   const {
     data,
@@ -85,6 +106,39 @@ export const PipelineRunProvider = ({
   useEffect(() => {
     fetchRunMetadata();
   }, [fetchRunMetadata]);
+
+  const rerun = useCallback(
+    async (
+      componentSpec: ComponentSpec,
+      options?: {
+        onSuccess?: (data: PipelineRun) => void;
+        onError?: (error: Error | string) => void;
+      },
+    ) => {
+      setIsSubmitting(true);
+      setError(null);
+      const authorizationRequired = isAuthorizationRequired();
+      if (authorizationRequired && !isAuthorized) {
+        const token = await awaitAuthorization();
+        if (token) {
+          authorizationToken.current = token;
+        }
+      }
+      await submitPipelineRun(componentSpec, backendUrl, {
+        authorizationToken: authorizationToken.current,
+        onSuccess: (data) => {
+          setIsSubmitting(false);
+          options?.onSuccess?.(data);
+        },
+        onError: (error) => {
+          setIsSubmitting(false);
+          setError(error);
+          options?.onError?.(error);
+        },
+      });
+    },
+    [isAuthorized, awaitAuthorization, backendUrl],
+  );
 
   useEffect(() => {
     if (executionError) {
@@ -122,9 +176,11 @@ export const PipelineRunProvider = ({
       status,
       metadata,
       isLoading,
+      isSubmitting,
       error,
+      rerun,
     }),
-    [details, state, status, metadata, isLoading, error],
+    [details, state, status, metadata, isLoading, isSubmitting, error, rerun],
   );
 
   return (
